chore: removes support for k0s
