+++ Justfile
@@ -73,10 +73,6 @@ generate-matrix-specific-images version="0.0.0":
     done
   done
 
-# Generate the CLI docs
-generate-cli-docs:
-  go run -mod vendor -tags pro ./hack/docs/main.go
-
 # Generate the vcluster.yaml config schema
 generate-config-schema:
   go run -mod vendor ./hack/schema/main.go
