+++ chart/tests/etcd-headless-service_test.yaml
@@ -37,35 +37,6 @@ tests:
           path: metadata.annotations.test
           value: test
 
-  - it: enable for k0s & defaults
-    release:
-      name: my-release
-      namespace: my-namespace
-    set:
-      controlPlane:
-        backingStore:
-          etcd:
-            deploy:
-              enabled: true
-              headlessService:
-                annotations:
-                  test: test
-        distro:
-          k0s:
-            enabled: true
-    asserts:
-      - hasDocuments:
-          count: 1
-      - equal:
-          path: metadata.name
-          value: my-release-etcd-headless
-      - equal:
-          path: metadata.namespace
-          value: my-namespace
-      - equal:
-          path: metadata.annotations.test
-          value: test
-
   - it: enable for k8s & defaults
     release:
       name: my-release
