+++ pkg/constants/constants.go
@@ -3,7 +3,6 @@ package constants
 const (
 	K8sKineEndpoint = "unix:///data/kine.sock"
 	K3sKineEndpoint = "unix:///data/server/kine.sock"
-	K0sKineEndpoint = "unix:///run/k0s/kine/kine.sock:2379"
 
 	K8sSqliteDatabase = "/data/state.db"
 	K3sSqliteDatabase = "/data/server/db/state.db"
