+++ pkg/etcd/client.go
@@ -51,13 +51,6 @@ func GetEtcdEndpoint(vConfig *config.VirtualClusterConfig) (string, *Certificate
 			ServerCert: "/data/pki/apiserver-etcd-client.crt",
 			ServerKey:  "/data/pki/apiserver-etcd-client.key",
 		}
-		if vConfig.Distro() == vconfig.K0SDistro {
-			etcdCertificates = &Certificates{
-				CaCert:     "/data/k0s/pki/etcd/ca.crt",
-				ServerCert: "/data/k0s/pki/apiserver-etcd-client.crt",
-				ServerKey:  "/data/k0s/pki/apiserver-etcd-client.key",
-			}
-		}
 
 		if vConfig.ControlPlane.BackingStore.Etcd.Embedded.Enabled {
 			etcdEndpoints = "https://127.0.0.1:2379"
@@ -77,18 +70,6 @@ func GetEtcdEndpoint(vConfig *config.VirtualClusterConfig) (string, *Certificate
 		etcdEndpoints = constants.K8sKineEndpoint
 	} else if vConfig.Distro() == vconfig.K3SDistro {
 		etcdEndpoints = constants.K3sKineEndpoint
-	} else if vConfig.Distro() == vconfig.K0SDistro {
-		if (vConfig.ControlPlane.BackingStore.Database.Embedded.Enabled && vConfig.ControlPlane.BackingStore.Database.Embedded.DataSource != "") ||
-			vConfig.ControlPlane.BackingStore.Database.External.Enabled {
-			etcdEndpoints = constants.K0sKineEndpoint
-		} else {
-			etcdEndpoints = "https://127.0.0.1:2379"
-			etcdCertificates = &Certificates{
-				CaCert:     "/data/k0s/pki/etcd/ca.crt",
-				ServerCert: "/data/k0s/pki/apiserver-etcd-client.crt",
-				ServerKey:  "/data/k0s/pki/apiserver-etcd-client.key",
-			}
-		}
 	}
 
 	return etcdEndpoints, etcdCertificates
